# docs/openapi.yml
openapi: 3.1.0
info:
  title: Tunecamp REST API
  version: 2.0.0
  description: |
    The Tunecamp API provides programmatic access to the platform's music library, artists, albums, tracks,
    releases, playlists, and administrative functions. It supports the decentralized music ecosystem via
    Zen and ActivityPub integrations.

    ## Authentication
    Public endpoints (catalog browsing, streaming) require no authentication.
    User endpoints require a Bearer token from `/api/auth/login`.
    Admin endpoints require a Bearer token with an admin role.

    ## Subsonic API
    For Subsonic/OpenSubsonic client compatibility, use the `/rest` endpoint.
    See [SUBSONIC.md](./SUBSONIC.md) for the full endpoint reference.
servers:
  - url: /api
    description: Tunecamp API base path
paths:
...
  /albums/{idOrSlug}:
    get:
      summary: Get Album Details
      tags: [Albums]
      description: Retrieves album details with track listing by ID or slug.
      operationId: getAlbum
      parameters:
        - name: idOrSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Album with tracks.
        '404':
          description: Album not found.
  /albums/{id}/cover:
    get:
      summary: Get Album Cover Art
      tags: [Albums]
      description: Returns the album's cover art image.
      operationId: getAlbumCover
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image file.

  # ── Tracks ──────────────────────────────────────────────────────────────────
  /tracks:
    get:
      summary: List Tracks
      tags: [Tracks]
      description: Retrieves all public tracks. Supports filtering by album.
      operationId: getTracks
      parameters:
        - name: albumId
          in: query
          schema:
            type: integer
          description: Filter tracks by album ID.
      responses:
        '200':
          description: A list of tracks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
  /tracks/{id}:
    get:
      summary: Get Track Details
      tags: [Tracks]
      description: Get a single track by ID.
      operationId: getTrack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Track details.
        '404':
          description: Track not found.
    put:
      summary: Update Track
      tags: [Tracks]
      description: Update track metadata (requires authentication, ownership check).
      operationId: updateTrack
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Track updated.
    delete:
      summary: Delete Track
      tags: [Tracks]
      description: Delete a track (requires authentication, ownership check).
      operationId: deleteTrack
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Track deleted.
  /tracks/{id}/stream:
    get:
      summary: Stream Track Audio
      tags: [Tracks]
      description: Stream the audio file for a track. Supports range requests.
      operationId: streamTrack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Audio stream.
        '404':
          description: Track not found.
  /tracks/{id}/cover:
    get:
      summary: Get Track Cover Art
      tags: [Tracks]
      description: Returns the cover art for a track (from album or embedded).
      operationId: getTrackCover
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image file.
  /tracks/batch:
    put:
      summary: Batch Update Tracks
      tags: [Tracks]
      description: Update metadata for multiple tracks at once.
      operationId: batchUpdateTracks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tracks updated.
    delete:
      summary: Batch Delete Tracks
      tags: [Tracks]
      description: Delete multiple tracks at once.
      operationId: batchDeleteTracks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tracks deleted.

  # ── Releases ────────────────────────────────────────────────────────────────
...
  /releases/{id}:
    get:
      summary: Get Release Details
      tags: [Releases]
      description: Get release details with track listing.
      operationId: getRelease
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Release with tracks.
    put:
      summary: Update Release
      tags: [Releases]
      description: Update a release (requires ownership or admin).
      operationId: updateRelease
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Release updated.
    delete:
      summary: Delete Release
      tags: [Releases]
      description: Delete a release (requires ownership or admin).
      operationId: deleteRelease
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Release deleted.

  # ── Playlists ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        album_id:
          type: integer
          nullable: true
        album_title:
          type: string
        artist_id:
          type: integer
          nullable: true
        artist_name:
          type: string
        track_num:
          type: integer
          nullable: true
        duration:
          type: number
          nullable: true
          description: Duration in seconds.
        file_path:
          type: string
          nullable: true
        format:
          type: string
          nullable: true
        bitrate:
          type: integer
          nullable: true
        sample_rate:
          type: integer
          nullable: true
        price:
          type: number
          nullable: true
        price_usdc:
          type: number
          nullable: true
        currency:
          type: string
          enum: [ETH, USD]
        lyrics:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...