# src/server/routes/tracks.ts
import { Router } from "express";
import fs from "fs-extra";
import path from "path";
import { parseFile } from "music-metadata";
import ffmpeg from "fluent-ffmpeg";
import ffmpegPath from "ffmpeg-static";
import type { DatabaseService } from "../database.js";
import type { LibraryService } from "../services/library.service.js";
import type { AuthenticatedRequest } from "../middleware/auth.js";
import { wrapAsync } from "../middleware/error-handling.js";
import { BadRequestError, ForbiddenError, NotFoundError } from "../common/errors.js";

// Set ffmpeg path
if (ffmpegPath) {
    ffmpeg.setFfmpegPath(ffmpegPath);
}

import type { AuthService } from "../auth.js";
import type { PublishingService } from "../publishing.js";
import { metadataService } from "../metadata.js";
import { VisibilityGuardian, Capability } from "../common/visibility.js";

export function createTracksRoutes(database: DatabaseService, publishingService: PublishingService, libraryService: LibraryService, musicDir: string, authService?: AuthService): Router {
    const router = Router();

    /**
     * GET /api/tracks
     * List all tracks
     */
    router.get("/", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        const showMine = req.query.mine === 'true';
        const tracks = await libraryService.getTracksForUser(
            { 
                userId: req.userId, 
                artistId: req.artistId,
                role: req.role, 
                isActive: req.isActive,
                username: req.username 
            },
            { mineOnly: showMine }
        );
        res.json(tracks);
    }));

    /**
     * GET /api/tracks/pricing/batch
     * Get pricing data for all tracks owned by the current user
     */
    router.get("/pricing/batch", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.isSuperUser && !req.artistId) throw new ForbiddenError("Unauthorized");

        const isRoot = req.isRootAdmin || req.isSuperUser;
        let tracksToSync: any[] = [];
        
        if (isRoot) {
           tracksToSync = database.getTracks();
        } else if (req.artistId) {
           tracksToSync = database.getTracksByOwner(req.artistId);
        } else {
           return res.json([]);
        }

        const artistIdsToFetch = [...new Set(tracksToSync
            .filter(t => t.price && t.price > 0 && t.artist_id)
            .map(t => t.artist_id as number))];

        const artistsBatch = database.getArtistsByIds(artistIdsToFetch);
        const artistMap = new Map(artistsBatch.map(a => [a.id, a]));

        const pricingData = tracksToSync
            .filter(t => t.price && t.price > 0)
            .map(t => ({
                trackId: t.id,
                price: t.price,
                currency: t.currency || 'ETH',
                priceUSDC: t.price_usdc || 0,
                priceUSDT: t.price_usdt || 0,
                walletAddress: t.artist_id ? artistMap.get(t.artist_id)?.wallet_address : null
            }));

        res.json(pricingData);
    }));

    /**
     * POST /api/tracks
     * Create a new track (usually for external links)
     */
    router.post("/", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.artistId) throw new ForbiddenError("Unauthorized");
        if (!req.isAdmin && !req.isActive) throw new ForbiddenError("Account not active");

        const { title, albumId, artistId: bodyArtistId, trackNum, url, service, externalArtwork, duration, lyrics, currency, priceUsdc } = req.body;
        
        let finalArtistId = bodyArtistId;
        if (!req.isAdmin) {
            finalArtistId = req.artistId;
        } else if (!req.isRootAdmin) {
            if (req.artistId) finalArtistId = req.artistId;
        }

        if (!title) throw new BadRequestError("Title is required");

        const trackId = database.createTrack({
            title,
            album_id: albumId || null,
            artist_id: finalArtistId || null,
            owner_id: req.userId || null,
            track_num: trackNum || null,
            duration: duration || 0,
            file_path: null, format: null, bitrate: null, sample_rate: null, lossless_path: null,
            url: url || null, service: service || null,
            external_artwork: externalArtwork || null,
            price: 0,
            price_usdc: priceUsdc !== undefined ? parseFloat(priceUsdc) : 0,
            currency: currency || 'ETH',
            waveform: null, lyrics: lyrics || null
        });

        const newTrack = database.getTrack(trackId);
        res.status(201).json(newTrack ? libraryService.mapTrackDTO(newTrack, req.username) : null);

        if (albumId) {
            publishingService.syncRelease(albumId).catch(e => console.error("Sync failed:", e));
        }
    }));

    /**
     * GET /api/tracks/search-metadata
     */
    router.get("/search-metadata", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.artistId) throw new ForbiddenError("Unauthorized");
        const query = req.query.q as string;
        if (!query) throw new BadRequestError("Query required");
        res.json(await metadataService.searchRecording(query));
    }));

    /**
     * PUT /api/tracks/batch
     */
    router.put("/batch", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.artistId) throw new ForbiddenError("Unauthorized");
        const { trackIds, data } = req.body;
        if (!Array.isArray(trackIds) || trackIds.length === 0) throw new BadRequestError("trackIds required");
        
        const isRoot = req.isRootAdmin;
        const results = await libraryService.batchUpdateTracks(trackIds, data, {
            userId: isRoot ? undefined : (req.userId ?? undefined),
            artistId: isRoot ? undefined : (req.artistId ?? undefined),
            isAdmin: !!req.isAdmin,
            username: req.username
        });
        res.json({ message: "Batch update completed", ...results });
    }));

    /**
     * DELETE /api/tracks/batch
     */
    router.delete("/batch", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.artistId) throw new ForbiddenError("Unauthorized");
        const { trackIds, deleteFiles } = req.body;
        if (!Array.isArray(trackIds) || trackIds.length === 0) throw new BadRequestError("trackIds required");
        
        const isRoot = req.isRootAdmin;
        const results = await libraryService.batchDeleteTracks(trackIds, deleteFiles === true, {
            userId: isRoot ? undefined : (req.userId ?? undefined),
            artistId: isRoot ? undefined : (req.artistId ?? undefined),
            isAdmin: !!req.isAdmin
        });
        res.json({ message: "Batch deletion completed", ...results });
    }));

    /**
     * GET /api/tracks/:id
     * Get track details
     */
    router.get("/:id", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        const id = parseInt(req.params.id as string, 10);
        if (isNaN(id)) throw new BadRequestError("Invalid track ID");

        const track = database.getTrack(id);
        if (!track) throw new NotFoundError("Track not found");

        const canSeePrivate = VisibilityGuardian.can(req.context, Capability.VIEW_PRIVATE_LIBRARY);
        if (!canSeePrivate && track.album_id) {
            const album = database.getAlbum(track.album_id);
            if (album && album.visibility === 'private' && track.owner_id !== req.userId) {
                if (!database.isTrackInPublicPlaylist(id)) throw new ForbiddenError("Access denied");
            }
        }

        res.json(libraryService.mapTrackDTO(track, req.username));
    }));

    /**
     * PUT /api/tracks/:id
     * Update track metadata and files
     */
    router.put("/:id", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.artistId) throw new ForbiddenError("Unauthorized");
        if (!req.isAdmin && !req.isActive) throw new ForbiddenError("Account not active");

        const id = parseInt(req.params.id as string, 10);
        const track = database.getTrack(id);
        if (!track) throw new NotFoundError("Track not found");

        const isRoot = req.isRootAdmin;
        const isOwner = isRoot || track.owner_id === req.userId || (track.owner_id === null && track.artist_id === req.artistId);
        if (!isRoot && !isOwner) throw new ForbiddenError("Access denied: You can only edit your own tracks");

        const updated = await libraryService.updateTrack(id, req.body);
        res.json({ message: "Track updated", track: updated ? libraryService.mapTrackDTO(updated, req.username) : null });
    }));

    /**
     * DELETE /api/tracks/:id
     * Delete track
     */
    router.delete("/:id", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.artistId) throw new ForbiddenError("Unauthorized");
        if (!req.isAdmin && !req.isActive) throw new ForbiddenError("Account not active");

        const id = parseInt(req.params.id as string, 10);
        const track = database.getTrack(id);
        if (!track) throw new NotFoundError("Track not found");

        const isRoot = req.isRootAdmin;
        const isOwner = isRoot || track.owner_id === req.userId || (track.owner_id === null && track.artist_id === req.artistId);
        if (!isRoot && !isOwner) throw new ForbiddenError("Access denied: You can only delete your own tracks");

        const deleteFile = req.query.deleteFile === "true";
        await libraryService.deleteTrack(id, deleteFile);
        res.json({ message: "Track deleted" });
    }));

    /**
     * POST /api/tracks/:id/star
     */
    router.post("/:id/star", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.username) throw new ForbiddenError("Unauthorized");
        const trackId = parseInt(req.params.id as string, 10);
        if (isNaN(trackId)) throw new BadRequestError("Invalid track ID");
        await libraryService.starTrack(req.username, trackId);
        res.json({ success: true, starred: true });
    }));

    /**
     * DELETE /api/tracks/:id/star
     */
    router.delete("/:id/star", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.username) throw new ForbiddenError("Unauthorized");
        const trackId = parseInt(req.params.id as string, 10);
        if (isNaN(trackId)) throw new BadRequestError("Invalid track ID");
        await libraryService.unstarTrack(req.username, trackId);
        res.json({ success: true, starred: false });
    }));

    /**
     * POST /api/tracks/:id/rating
     */
    router.post("/:id/rating", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.username) throw new ForbiddenError("Unauthorized");
        const trackId = parseInt(req.params.id as string, 10);
        const { rating } = req.body;
        const r = parseInt(rating);
        if (isNaN(r) || r < 0 || r > 5) throw new BadRequestError("Invalid rating (must be 0-5)");
        await libraryService.setTrackRating(req.username, trackId, r);
        res.json({ success: true, rating: r });
    }));

    /**
     * GET /api/tracks/:id/lyrics
     */
    router.get("/:id/lyrics", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        const id = parseInt(req.params.id as string, 10);
        const track = database.getTrack(id);
        if (!track || !track.file_path) throw new NotFoundError("Track not found");

        const trackPath = path.join(musicDir, track.file_path);
        if (!await fs.pathExists(trackPath)) throw new NotFoundError("File not found");

        const metadata = await parseFile(trackPath).catch(() => null);
        let lyrics = track.lyrics || "";
        if (!lyrics && metadata?.common?.lyrics) {
            const l = metadata.common.lyrics;
            lyrics = Array.isArray(l) && l.length > 0 ? (typeof l[0] === 'string' ? l[0] : (l[0] as any).text || "") : (typeof l === 'string' ? l : "");
        }
        res.json({ lyrics });
    }));

    /**
     * GET /api/tracks/:id/cover
     */
    router.get("/:id/cover", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        const id = parseInt(req.params.id as string, 10);
        const track = database.getTrack(id);
        if (!track) throw new NotFoundError("Track not found");

        if (track.external_artwork) {
            if (track.external_artwork.startsWith('http')) return res.redirect(track.external_artwork);
            const artworkPath = path.join(musicDir, track.external_artwork);
            if (await fs.pathExists(artworkPath)) return res.sendFile(path.resolve(artworkPath), { maxAge: 86400000 });
        }

        if (track.album_id) return res.redirect(`/api/albums/${track.album_id}/cover`);

        const { getPlaceholderSVG } = await import("../../utils/audioUtils.js");
        const svg = getPlaceholderSVG(track.title || "No Cover");
        res.setHeader("Content-Type", "image/svg+xml").setHeader("Cache-Control", "public, max-age=3600");
        return res.send(svg);
    }));

    /**
     * GET /api/tracks/:id/metadata (extract from file)
     */
    router.get("/:id/metadata", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.artistId) throw new ForbiddenError("Unauthorized");
        const id = parseInt(req.params.id as string, 10);
        const track = database.getTrack(id);
        if (!track || !track.file_path) throw new NotFoundError("Track not found");

        const trackPath = path.join(musicDir, track.file_path);
        const metadata = await parseFile(trackPath).catch(() => null);
        if (!metadata) throw new Error("Failed to parse metadata");

        const { common } = metadata;
        let coverBase64 = null;
        if (common.picture && common.picture.length > 0) {
            const pic = common.picture[0];
            coverBase64 = `data:${pic.format};base64,${Buffer.from(pic.data).toString('base64')}`;
        }

        res.json({
            title: common.title || track.title,
            artist: common.artist || common.albumartist,
            album: common.album,
            year: common.year,
            genre: common.genre ? common.genre.join(", ") : undefined,
            cover: coverBase64
        });
    }));



    /**
     * POST /api/tracks/:id/match-metadata
     */
    router.post("/:id/match-metadata", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        if (!req.isAdmin && !req.artistId) throw new ForbiddenError("Unauthorized");
        const id = parseInt(req.params.id, 10);
        const track = database.getTrack(id);
        if (!track) throw new NotFoundError("Track not found");
        
        const isRoot = req.isRootAdmin;
        if (!isRoot && !req.isAdmin && track.owner_id !== req.artistId) throw new ForbiddenError("Access denied");

        const { title, artist, albumTitle, coverUrl } = req.body;
        console.log(`🔍 [Metadata Match] Processing track ${id}:`, { title, artist, albumTitle, hasCover: !!coverUrl });

        try {
            let artistId = track.artist_id;
            if (artist) {
                console.log(`   - Matching artist: "${artist}"`);
                const a = database.getArtistByName(artist);
                artistId = a ? a.id : database.createArtist(artist);
                console.log(`   - Artist ID: ${artistId}`);
                database.updateTrackArtist(id, artistId);
            }

            if (albumTitle) {
                console.log(`   - Matching album: "${albumTitle}"`);
                const slug = "lib-" + albumTitle.toLowerCase().replace(/[^a-z0-9]/g, '-');
                let alb = database.getAlbumBySlug(slug);
                if (!alb) {
                    console.log(`   - Creating new library album: "${albumTitle}" (slug: ${slug})`);
                    const newId = database.createAlbum({
                        title: albumTitle, slug, artist_id: artistId, owner_id: req.userId || null,
                        date: null, cover_path: null, genre: "Matched", description: "Matched",
                        type: 'album', year: null, download: null, price: 0, price_usdc: 0, currency: 'ETH',
                        external_links: null, is_public: false, visibility: 'private', is_release: false,
                        published_at: null, published_to_gundb: false, published_to_ap: false, license: null,
                        status: 'draft',
                    });
                    alb = database.getAlbum(newId);
                }
                if (alb) {
                    console.log(`   - Associating track with album ID: ${alb.id}`);
                    database.updateTrackAlbum(id, alb.id);
                }
            }

            if (title) {
                console.log(`   - Updating title: "${title}"`);
                database.updateTrackTitle(id, title);
            }
            
            if (coverUrl) {
                console.log(`   - Updating external artwork: ${coverUrl.substring(0, 50)}...`);
                (database as any).db.prepare("UPDATE tracks SET external_artwork = ? WHERE id = ?").run(coverUrl, id);
            }

            const updated = database.getTrack(id);
            res.json({ message: "Metadata matched", track: updated ? libraryService.mapTrackDTO(updated, req.username) : null });
        } catch (error: any) {
            console.error(`❌ [Metadata Match Error] Track ${id}:`, error);
            throw error;
        }
    }));

    /**
     * GET /api/tracks/:id/stream
     */
    router.get("/:id/stream", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        const id = parseInt(req.params.id as string, 10);
        const track = database.getTrack(id);
        if (!track) throw new NotFoundError("Track not found");

        const isOwner = (req.userId !== undefined && track.owner_id === req.userId) || (req.artistId !== undefined && track.artist_id === req.artistId);
        const canSeePrivate = VisibilityGuardian.can(req.context, Capability.VIEW_PRIVATE_LIBRARY);
        if (!canSeePrivate && !isOwner) {
            if (track.album_id) {
                const album = database.getRelease(track.album_id) || database.getAlbum(track.album_id);
                if (album && album.visibility === 'private' && !database.isTrackInPublicPlaylist(id)) throw new ForbiddenError("Access denied");
            } else throw new ForbiddenError("Access denied");
        }

        if (!track.file_path) throw new NotFoundError("Track file not found");
        let trackPath = path.join(musicDir, track.file_path);
        let usingLosslessFallback = false;

        if (!await fs.pathExists(trackPath)) {
            const decoded = decodeURIComponent(trackPath);
            if (await fs.pathExists(decoded)) trackPath = decoded;
            else if (track.lossless_path) {
                let lp = path.join(musicDir, track.lossless_path);
                if (!await fs.pathExists(lp)) lp = decodeURIComponent(lp);
                if (await fs.pathExists(lp)) { trackPath = lp; usingLosslessFallback = true; }
                else throw new NotFoundError("Audio file not found");
            } else throw new NotFoundError("Audio file not found");
        }

        const stat = await fs.promises.stat(trackPath);
        const ext = path.extname(trackPath).toLowerCase();
        let targetFormat = req.query.format as string;
        
        // Only force MP3 transcoding for WAV (heavy/poorly supported) or when using a lossless fallback
        // FLAC is well-supported natively and should be served as-is for better seeking/quality
        if (!targetFormat && (ext === '.wav' || usingLosslessFallback)) targetFormat = 'mp3';

        if (targetFormat && (targetFormat !== ext.substring(1) || usingLosslessFallback)) {
            const contentTypeMap: any = { 'mp3': 'audio/mpeg', 'aac': 'audio/aac', 'ogg': 'audio/ogg', 'opus': 'audio/opus' };
            res.setHeader("Content-Type", contentTypeMap[targetFormat] || 'audio/mpeg');
            
            ffmpeg(trackPath)
                .format(targetFormat)
                .audioCodec(targetFormat === 'mp3' ? 'libmp3lame' : 'copy')
                .audioChannels(2)
                .audioBitrate((req.query.bitrate as string) || '192k')
                .on('error', (err) => { 
                    if (!err.message.includes("Output stream closed")) {
                        console.error('Transcoding error:', err.message);
                    }
                })
                .pipe(res, { end: true });
            return;
        }

        const range = req.headers.range;
        const contentTypes: any = { ".mp3": "audio/mpeg", ".flac": "audio/flac", ".ogg": "audio/ogg", ".wav": "audio/wav", ".m4a": "audio/mp4", ".aac": "audio/aac", ".opus": "audio/opus" };
        const contentType = contentTypes[ext] || "audio/mpeg";

        if (range) {
            const parts = range.replace(/bytes=/, "").split("-");
            const start = parseInt(parts[0], 10);
            const end = parts[1] ? parseInt(parts[1], 10) : stat.size - 1;
            res.writeHead(206, { "Content-Range": `bytes ${start}-${end}/${stat.size}`, "Accept-Ranges": "bytes", "Content-Length": end - start + 1, "Content-Type": contentType });
            fs.createReadStream(trackPath, { start, end }).pipe(res);
        } else {
            res.writeHead(200, { "Content-Length": stat.size, "Content-Type": contentType, "Accept-Ranges": "bytes" });
            fs.createReadStream(trackPath).pipe(res);
        }
    }));

    /**
     * GET /api/tracks/:id/download
     * Download original file
     */
    router.get("/:id/download", wrapAsync(async (req: AuthenticatedRequest, res: any) => {
        const id = parseInt(req.params.id as string, 10);
        const track = database.getTrack(id);
        if (!track) throw new NotFoundError("Track not found");

        const isOwner = (req.userId !== undefined && track.owner_id === req.userId) || (req.artistId !== undefined && track.artist_id === req.artistId);
        const canSeePrivate = VisibilityGuardian.can(req.context, Capability.VIEW_PRIVATE_LIBRARY);
        if (!canSeePrivate && !isOwner) {
            if (track.album_id) {
                const album = database.getRelease(track.album_id) || database.getAlbum(track.album_id);
                if (album && album.visibility === 'private' && !database.isTrackInPublicPlaylist(id)) throw new ForbiddenError("Access denied");
            } else throw new ForbiddenError("Access denied");
        }

        if (!track.file_path) throw new NotFoundError("Track file not found");
        let trackPath = path.join(musicDir, track.file_path);

        if (!await fs.pathExists(trackPath)) {
            const decoded = decodeURIComponent(trackPath);
            if (await fs.pathExists(decoded)) trackPath = decoded;
            else if (track.lossless_path) {
                let lp = path.join(musicDir, track.lossless_path);
                if (!await fs.pathExists(lp)) lp = decodeURIComponent(lp);
                if (await fs.pathExists(lp)) trackPath = lp;
                else throw new NotFoundError("Audio file not found");
            } else throw new NotFoundError("Audio file not found");
        }

        const ext = path.extname(trackPath);
        const filename = `${track.artist_name || 'Unknown'} - ${track.title || 'Untitled'}${ext}`;
        
        res.download(trackPath, filename);
    }));

    return router;
}